+++ tests/conftest.py
@@ -1,4 +1,5 @@
 from io import StringIO
+from typing import AsyncIterable, List
 
 import pytest
 from faker import Faker
@@ -25,3 +26,16 @@ def error_buffer() -> StringIO:
 @pytest.fixture()
 def cli_output(stream_buffer: StringIO, error_buffer: StringIO) -> CLIOutput:
     return CLIOutput(stream_buffer, error_buffer)
+
+
+@pytest.fixture()
+def random_numbers(faker: Faker) -> List[int]:
+    return faker.pylist(faker.pyint(10, 80), False, [int])
+
+
+@pytest.fixture()
+def random_number_async_iterable(faker: Faker, random_numbers: List[int]) -> AsyncIterable:
+    async def generator():
+        for number in random_numbers:
+            yield number
+    return generator()